add custom name feature
Disappointingly, jsdom doesn't seem to support even client-side
navigation, so the tests for navigating to the search feature are based
on providing an initial route rather than clicking the search button.
With more time I'd add some tests to assert that the button is at least
requesting the correct URL change.